obs, net: Deduplicate update_nad_references into libs/net/vmspec
Move update_nad_references to libs/net/vmspec.py to eliminate identical
copies in tests/utils.py and tests/network/l2_bridge/nad_ref_change/lib_helpers.py.
The function is used by both network and observability team directories,
so it belongs in a shared module per project conventions.

assisted by: claude code claude-opus-4-6
