+++ libs/net/vmspec.py
@@ -1,5 +1,6 @@
 import ipaddress
 from collections.abc import Callable
+from copy import deepcopy
 from typing import Any, Final
 
 from kubernetes.dynamic.client import ResourceField
@@ -299,3 +300,23 @@ def _vmi_condition_not_set(existing_conditions: list[ResourceField], required_co
         for cond in existing_conditions
         if cond.type == required_condition
     )
+
+
+def update_nad_references(vm: BaseVirtualMachine, nad_name_by_net: dict[str, str]) -> None:
+    """Update secondary network NAD references and wait for the change to be fully applied.
+
+    Patches the VM spec atomically, then waits for the MigrationRequired condition to
+    appear (change detected) and disappear (migration completed).
+
+    Args:
+        vm: The virtual machine to update.
+        nad_name_by_net: Mapping of interface name to new NAD name.
+    """
+    resource_version = vm.vmi.instance.metadata.resourceVersion
+    networks = deepcopy(vm.template_spec.networks) or []
+    for network in networks:
+        if network.name in nad_name_by_net and network.multus:
+            network.multus.networkName = nad_name_by_net[network.name]
+    vm.set_networks(networks=networks)
+    wait_for_vmi_condition_status(vm=vm, condition="MigrationRequired", resource_version=resource_version)
+    wait_for_no_vmi_condition(vm=vm, condition="MigrationRequired")
