+++ tests/utils.py
@@ -5,7 +5,6 @@ import re
 import shlex
 import tarfile
 from contextlib import contextmanager
-from copy import deepcopy
 from io import BytesIO
 from typing import Generator, Optional
 
@@ -27,8 +26,6 @@ from pyhelper_utils.shell import run_ssh_commands
 from pytest_testconfig import config as py_config
 from timeout_sampler import TimeoutExpiredError, TimeoutSampler, retry
 
-from libs.net.vmspec import wait_for_no_vmi_condition, wait_for_vmi_condition_status
-from libs.vm.vm import BaseVirtualMachine
 from utilities.artifactory import (
     cleanup_artifactory_secret_and_config_map,
     get_artifactory_config_map,
@@ -782,23 +779,3 @@ def create_windows2022_vm_with_vtpm_from_registry(
         running_vm(vm=vm)
         wait_for_windows_vm(vm=vm, version="2022")
         yield vm
-
-
-def update_nad_references(vm: BaseVirtualMachine, nad_name_by_net: dict[str, str]) -> None:
-    """Update secondary network NAD references and wait for the change to be fully applied.
-
-    Patches the VM spec atomically, then waits for the MigrationRequired condition to
-    appear (change detected) and disappear (migration completed).
-
-    Args:
-        vm: The virtual machine to update.
-        nad_name_by_net: Mapping of interface name to new NAD name.
-    """
-    resource_version = vm.vmi.instance.metadata.resourceVersion
-    networks = deepcopy(vm.template_spec.networks) or []
-    for network in networks:
-        if network.name in nad_name_by_net and network.multus:
-            network.multus.networkName = nad_name_by_net[network.name]
-    vm.set_networks(networks=networks)
-    wait_for_vmi_condition_status(vm=vm, condition="MigrationRequired", resource_version=resource_version)
-    wait_for_no_vmi_condition(vm=vm, condition="MigrationRequired")
